vfs: Fix counting in `inode_avail_count`
As seen in `inode_del` (and `inode_alloc`), a free (`NOT_INIT`) node can be found anywhere in `g_vfs_dev_nodes`; it’s a “sparse” list.
So when checking for free nodes, the iterating index variable (`e`) shall be used, not the counter (`count`). The code got that wrong.
For instance, if the first node is not available (type is not `VFS_TYPE_NOT_INIT`), then the function previously returned 0, no matter what, and was iterating the node list in vain.